Moved specific scalar type implementations to individual folders for sanity
