+++ scalars/src/scalar_implementations/half/half_random_generator.h
@@ -1,7 +1,7 @@
-// Copyright (c) 2023 Datasig Developers. All rights reserved.
+// Copyright (c) 2023 the RoughPy Developers. All rights reserved.
 //
-// Redistribution and use in source and binary forms, with or without
-// modification, are permitted provided that the following conditions are met:
+// Redistribution and use in source and binary forms, with or without modification,
+// are permitted provided that the following conditions are met:
 //
 // 1. Redistributions of source code must retain the above copyright notice,
 // this list of conditions and the following disclaimer.
@@ -18,13 +18,12 @@
 // AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
 // IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
 // ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
-// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-// POSSIBILITY OF SUCH DAMAGE.
+// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+// DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+// SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+// CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+// OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
+// USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 //
 // Created by user on 26/04/23.
@@ -33,27 +32,27 @@
 #ifndef ROUGHPY_SCALARS_SRC_HALF_RANDOM_GENERATOR_H
 #define ROUGHPY_SCALARS_SRC_HALF_RANDOM_GENERATOR_H
 
-#include "random_impl.h"
-#include <roughpy/scalars/random.h>
-#include <roughpy/scalars/scalar.h>
-#include <roughpy/scalars/scalar_type.h>
+#include "../../random_impl.h"
+#include "random.h"
+#include "scalar.h"
+#include "scalar_type.h"
 
-#include "standard_random_generator.h"
+#include "../../standard_random_generator.h"
 
 #include <mutex>
 #include <random>
+#include <sstream>
 #include <vector>
 
 #include <pcg_random.hpp>
-#include <sstream>
 
 namespace rpy {
 namespace scalars {
 
 template <typename BitGenerator>
-class StandardRandomGenerator<bfloat16, BitGenerator> : public RandomGenerator
+class StandardRandomGenerator<half, BitGenerator> : public RandomGenerator
 {
-    using scalar_type = bfloat16;
+    using scalar_type = half;
     using bit_generator = BitGenerator;
 
     std::vector<uint64_t> m_seed;
@@ -71,6 +70,7 @@ public:
             = delete;
 
     void set_seed(Slice<uint64_t> seed_data) override;
+    void set_state(string_view state) override;
     std::vector<uint64_t> get_seed() const override;
     string get_type() const override;
     string get_state() const override;
@@ -79,24 +79,24 @@ public:
                                            dimn_t count) const override;
     OwnedScalarArray normal_random(Scalar loc, Scalar scale,
                                    dimn_t count) const override;
-    void set_state(string_view state) override;
 };
 
 template <typename BitGenerator>
-string StandardRandomGenerator<bfloat16, BitGenerator>::get_state() const
+void StandardRandomGenerator<half, BitGenerator>::set_state(string_view state)
+{
+    std::stringstream ss(string{state});
+    ss >> m_generator;
+}
+template <typename BitGenerator>
+string StandardRandomGenerator<half, BitGenerator>::get_state() const
 {
     std::stringstream ss;
     ss << m_generator;
     return ss.str();
 }
-template <typename BitGenerator>
-string StandardRandomGenerator<bfloat16, BitGenerator>::get_type() const
-{
-    return std::string(dtl::rng_type_getter<BitGenerator>::name);
-}
 
 template <typename BitGenerator>
-StandardRandomGenerator<bfloat16, BitGenerator>::StandardRandomGenerator(
+StandardRandomGenerator<half, BitGenerator>::StandardRandomGenerator(
         const ScalarType* stype, Slice<uint64_t> seed)
     : RandomGenerator(stype), m_seed{seed[0]},
       m_generator(BitGenerator(seed[0]))
@@ -106,7 +106,7 @@ StandardRandomGenerator<bfloat16, BitGenerator>::StandardRandomGenerator(
 }
 
 template <typename BitGenerator>
-void StandardRandomGenerator<bfloat16, BitGenerator>::set_seed(
+void StandardRandomGenerator<half, BitGenerator>::set_seed(
         Slice<uint64_t> seed_data)
 {
     RPY_CHECK(seed_data.size() >= 1);
@@ -114,21 +114,20 @@ void StandardRandomGenerator<bfloat16, BitGenerator>::set_seed(
     m_seed = {seed_data[0]};
 }
 template <typename BitGenerator>
-void StandardRandomGenerator<bfloat16, BitGenerator>::set_state(
-        string_view state)
-{
-    std::stringstream ss(string{state});
-    ss >> m_generator;
-}
-template <typename BitGenerator>
 std::vector<uint64_t>
-StandardRandomGenerator<bfloat16, BitGenerator>::get_seed() const
+StandardRandomGenerator<half, BitGenerator>::get_seed() const
 {
     return {m_seed[0]};
 }
 template <typename BitGenerator>
+std::string StandardRandomGenerator<half, BitGenerator>::get_type() const
+{
+    return std::string(dtl::rng_type_getter<BitGenerator>::name);
+}
+
+template <typename BitGenerator>
 OwnedScalarArray
-StandardRandomGenerator<bfloat16, BitGenerator>::uniform_random_scalar(
+StandardRandomGenerator<half, BitGenerator>::uniform_random_scalar(
         ScalarArray lower, ScalarArray upper, dimn_t count) const
 {
     std::vector<std::uniform_real_distribution<float>> dists;
@@ -137,40 +136,39 @@ StandardRandomGenerator<bfloat16, BitGenerator>::uniform_random_scalar(
 
     dists.reserve(lower.size());
     for (dimn_t i = 0; i < lower.size(); ++i) {
-        dists.emplace_back(static_cast<float>(scalar_cast<bfloat16>(lower[i])),
-                           static_cast<float>(scalar_cast<bfloat16>(upper[i])));
+        dists.emplace_back(static_cast<float>(scalar_cast<half>(lower[i])),
+                           static_cast<float>(scalar_cast<half>(upper[i])));
     }
 
     OwnedScalarArray result(p_type, count * dists.size());
 
-    auto* out = result.raw_cast<bfloat16*>();
+    auto* out = result.raw_cast<half*>();
     for (dimn_t i = 0; i < count; ++i) {
-        for (auto& dist : dists) { ::new (out++) bfloat16(dist(m_generator)); }
+        for (auto& dist : dists) { ::new (out++) half(dist(m_generator)); }
     }
 
     return result;
 }
 template <typename BitGenerator>
-OwnedScalarArray StandardRandomGenerator<bfloat16, BitGenerator>::normal_random(
+OwnedScalarArray StandardRandomGenerator<half, BitGenerator>::normal_random(
         Scalar loc, Scalar scale, dimn_t count) const
 {
 
     OwnedScalarArray result(p_type, count);
-    std::normal_distribution<float> dist(
-            static_cast<float>(scalar_cast<bfloat16>(loc)),
-            static_cast<float>(scalar_cast<bfloat16>(scale)));
+    std::normal_distribution<float> dist(scalar_cast<half>(loc),
+                                         scalar_cast<half>(scale));
 
-    auto* out = result.raw_cast<bfloat16*>();
+    auto* out = result.raw_cast<half*>();
     for (dimn_t i = 0; i < count; ++i) {
-        ::new (out++) bfloat16(dist(m_generator));
+        ::new (out++) half(dist(m_generator));
     }
 
     return result;
 }
 
-extern template class StandardRandomGenerator<bfloat16, std::mt19937_64>;
+extern template class StandardRandomGenerator<half, std::mt19937_64>;
 
-extern template class StandardRandomGenerator<bfloat16, pcg64>;
+extern template class StandardRandomGenerator<half, pcg64>;
 
 }// namespace scalars
 }// namespace rpy
